Repair Mathematica File
Modified mathematica file to reflect correct Possion PMF.  Additionly, changed FOC back to a solvable form.  The addion of the "A" to the exponent in the second term resulted memory use beyond that of any modern computer.  This restates the utility maximization such that A is a multiplier and not an exponent.