Updated Mathematica File After Failed Run
I grabbed the Mathematica files from the office simulation.  However, I noticed that it didn’t output anything.  It appears some bugs have entered into the program after our testing in March.  I’ve tried to correct these the best I can, but in one case I’m not positive what was intended.

(1) I’ve commented out the line with LogNormal as part of pfunc.  LogNormal defines a distribution, but it doesn’t define if it is a PDF/CDF and the value you wish to pull.  As of now, there is not enough values specified to pull a number.
(2) There was a series of undefined variables specified as shocks (e.g. Rt).  However, strings needed to be specified (e.g. "Rt").  Thus the simulations were run with a shock of "" (blank).  Thus nothing was produced.   This I believe is now fixed.